Auto-calculate reset threshold
What's changed?
When the user changes the activation threshold, the reset threshold is now automatically calculated to always be one degree lower than the activation threshold. And because the reset threshold is now auto-calculated, it's no longer necessary to be configurable to the user so has been removed from the usermod settings page to keep it simple.

Why the difference between the thresholds?
To prevent frequent toggling between states when the temperature is close to the threshold, the reset threshold is slightly lower than the activation threshold to provide a small buffer.